pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-photos-manager"
version = "2.1.0"
description = "Intelligent photo organization and duplicate management tool for Google Photos Takeout archives"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Google Photos Manager Contributors"}
]
keywords = [
    "photos",
    "google-photos",
    "takeout",
    "duplicate-detection",
    "photo-management",
    "organization",
    "android",
    "kivy"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]

dependencies = [
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
mobile = [
    "kivy>=2.2.1",
    "kivymd>=1.1.1",
    "buildozer>=1.5.0",
    "cython>=3.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-timeout>=2.1.0",
    "pytest-mock>=3.11.0",
    "pytest-benchmark>=4.0.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "bandit>=1.7.5",
    "pre-commit>=3.5.0",
    "coverage>=7.3.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/NickNak86/Mikes_GooglePhoto-Fixer"
Repository = "https://github.com/NickNak86/Mikes_GooglePhoto-Fixer"
Issues = "https://github.com/NickNak86/Mikes_GooglePhoto-Fixer/issues"
Changelog = "https://github.com/NickNak86/Mikes_GooglePhoto-Fixer/blob/main/CHANGELOG.md"
Documentation = "https://github.com/NickNak86/Mikes_GooglePhoto-Fixer#readme"

[project.scripts]
photo-manager = "PhotoManager:main"
photo-manager-mobile = "PhotoManagerMobile:main"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | \.buildozer
  | bin
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip_glob = [
    "**/.venv/*",
    "**/venv/*",
    "**/.buildozer/*",
    "**/bin/*",
    "**/build/*",
    "**/dist/*",
]

# mypy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
show_error_codes = true
show_column_numbers = true
pretty = true
exclude = [
    '\.venv',
    'venv',
    '\.buildozer',
    'bin',
    'build',
    'dist',
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "-W", "default",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "mobile: marks tests as mobile-specific",
    "desktop: marks tests as desktop-specific",
    "requires_opencv: marks tests that require OpenCV",
    "requires_tkinter: marks tests that require tkinter",
    "performance: marks tests as performance benchmarks",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
timeout = 300
timeout_method = "thread"

# coverage.py configuration
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/dist-packages/*",
    "*/venv/*",
    "*/.venv/*",
    "setup.py",
]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
    "pass",
    "\\.\\.\\.",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

# Bandit security linter configuration
[tool.bandit]
targets = ["."]
exclude_dirs = [
    "/tests/",
    "/.venv/",
    "/venv/",
    "/.buildozer/",
    "/bin/",
    "/build/",
    "/dist/",
]
skips = [
    "B101",  # assert_used - we use asserts in tests
    "B601",  # paramiko_calls - not using paramiko
]

# Interrogate docstring coverage configuration
[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 50
exclude = ["setup.py", "docs", "build", "tests"]
verbose = 1
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false

# setuptools configuration
[tool.setuptools]
package-dir = {"" = "."}
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = [
    "tests*",
    "docs*",
    "*.tests",
    "*.tests.*",
    "tests.*",
]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.json", "*.yml", "*.yaml"]
